requirements.txt
absl-py==1.4.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
efficientnet-pytorch==0.6.3
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.14.0
flatbuffers==23.5.26
fonttools==4.40.0
fqdn==1.5.1
fsspec==2024.3.1
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
huggingface-hub==0.22.2
idna==2.8
imageio==2.34.1
imgaug==0.4.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.24.0
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.4.2
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-http-over-ws==0.0.8
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
munch==4.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
nibabel==5.2.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.20.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.8.0
pretrainedmodels==0.7.4
prometheus-client==0.17.0
prompt-toolkit==3.0.39
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyGObject==3.36.0
pyparsing==3.0.9
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.29.1
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.8.4
rsa==4.9
safetensors==0.4.3
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.9.3
segmentation-models-pytorch==0.2.0
Send2Trash==1.8.2
shapely==2.0.4
SimpleITK==2.3.1
six==1.14.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.4.12
tinycss2==1.2.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.3.2
tqdm==4.66.2
traitlets==5.9.0
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==1.25.8
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.8
wrapt==1.15.0
zipp==3.15.0
